fix(heat): prevent arithmetic overflow

+++ Cores/SystHeats/HeatConductionModel/HeatConductionModel.cpp
@@ -417,7 +417,7 @@ tuple<vector<double>, vector<double>> HeatConductionModel::GenerateCoeAndSrcMatr
 
     int size = mGrid->GetNumCells();
 
-    vector<double> matrix(size * size, 0);
+    vector<double> matrix(static_cast<long>(size) * size, 0);
     vector<double> source(size, 0);
 
     for (int i = 0; i < size; i++)
